Refactored commands into separate files
