+++ src-tauri/src/ipc/commands/mod.rs
@@ -1,0 +1,4 @@
+pub mod connections;
+pub mod graph;
+pub mod mesh;
+pub mod radio;
